Update packages, refactor views, and improve UI
- Updated `Microsoft.Extensions.Logging.Debug` to `9.0.5` and added `Newtonsoft.Json` version `13.0.3`.
- Changed path for `AvatarSelectBottomSheet.xaml` and added new XAML files for controls.
- Removed unused popup references and their associated view models.
- Adjusted styles and namespaces for better organization.
- Enhanced `BowlerListOverviewViewModel` methods for improved data handling.
- Introduced new loading indicators and improved UI layout across various pages.
- Refactored `ContentPageBase` and updated related XAML files for consistency.
