--- HangTab/ViewModels/Popups/DataResetPopUpViewModel.cs
@@ -1,6 +1,0 @@
-﻿using CommunityToolkit.Maui.Core;
-
-namespace HangTab.ViewModels.Popups;
-public partial class DataResetPopUpViewModel(IPopupService popupService) : PopupViewModelBase(popupService)
-{
-}
