--- HangTab/ViewModels/Popups/DeleteBowlerPopupViewModel.cs
@@ -1,6 +1,0 @@
-﻿using CommunityToolkit.Maui.Core;
-
-namespace HangTab.ViewModels.Popups;
-public partial class DeleteBowlerPopupViewModel(IPopupService popupService) : PopupViewModelBase(popupService)
-{
-}
