--- HangTab/ViewModels/Popups/StartNewSeasonPopupViewModel.cs
@@ -1,6 +1,0 @@
-﻿using CommunityToolkit.Maui.Core;
-
-namespace HangTab.ViewModels.Popups;
-public partial class StartNewSeasonPopupViewModel(IPopupService popupService) : PopupViewModelBase(popupService)
-{
-}
