+++ HangTab/ViewModels/SettingsViewModel.cs
@@ -1,5 +1,4 @@
-﻿using CommunityToolkit.Maui.Core;
-using CommunityToolkit.Mvvm.ComponentModel;
+﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.Messaging;
 
@@ -7,13 +6,11 @@ using HangTab.Enums;
 using HangTab.Messages;
 using HangTab.Services;
 using HangTab.ViewModels.Base;
-using HangTab.ViewModels.Popups;
 
 namespace HangTab.ViewModels;
 public partial class SettingsViewModel(
     IDatabaseService databaseService,
     IDialogService dialogService,
-    IPopupService popupService,
     ISettingsService settingsService,
     IThemeService themeService) : ViewModelBase
 {
@@ -38,44 +35,30 @@ public partial class SettingsViewModel(
         }
     }
 
-    public override async Task LoadAsync()
-    {
-        await Loading(
-            () =>
-            {
-                TotalSeasonWeeks = settingsService.TotalSeasonWeeks;
-                DarkThemeEnabled = settingsService.Theme == (int)Theme.Dark;
-                return Task.CompletedTask;
-            });
-
-    }
+    public override async Task LoadAsync() => await Loading(InitializeSettingsAsync);
 
     [RelayCommand]
     private async Task DeleteAllDataAsync()
     {
-        var result = await popupService.ShowPopupAsync<DataResetPopUpViewModel>();
-        if (result is null)
+        if (await dialogService.Ask("Delete", "Are you sure you want to delete ALL data?", "Yes", "No"))
         {
-            return;
-        }
-
-        if (await databaseService.DeleteAllTableData())
-        {
-            settingsService.CurrentWeekPrimaryKey = 0;
-            WeakReferenceMessenger.Default.Send(new SystemResetMessage());
-            await dialogService.ToastAsync("All data has been deleted");
-        }
-        else
-        {
-            await dialogService.AlertAsync("Critical Error", "Error occurred while deleting the databases!", "Ok");
+            if (await databaseService.DeleteAllTableData())
+            {
+                settingsService.CurrentWeekPrimaryKey = 0;
+                WeakReferenceMessenger.Default.Send(new SystemResetMessage());
+                await dialogService.ToastAsync("All data has been deleted");
+            }
+            else
+            {
+                await dialogService.AlertAsync("Critical Error", "Error occurred while deleting the databases!", "Ok");
+            }
         }
     }
 
     [RelayCommand]
     private async Task StartNewSeasonAsync()
     {
-        var result = await popupService.ShowPopupAsync<StartNewSeasonPopupViewModel>();
-        if (result is null)
+        if (await dialogService.Ask("Reset", "Are you ready to start a new season and reset all bowler hangings?", "Yes", "No"))
         {
             return;
         }
@@ -95,5 +78,12 @@ public partial class SettingsViewModel(
             //    await dialogService.AlertAsync("Critical Error", "Error occurred while resetting databaseService!", "Ok");
             //}
         //}
+    }
+
+    private Task InitializeSettingsAsync()
+    {
+        TotalSeasonWeeks = settingsService.TotalSeasonWeeks;
+        DarkThemeEnabled = settingsService.Theme == (int)Theme.Dark;
+        return Task.CompletedTask;
     }
 }
