+++ HangTab/Views/BottomSheets/AvatarSelectBottomSheet.xaml
@@ -3,8 +3,8 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:the49="https://schemas.the49.com/dotnet/2023/maui"
-    xmlns:vm="clr-namespace:HangTab.ViewModels"
-    x:Class="HangTab.Views.AvatarSelectBottomSheet"
+    xmlns:vm="clr-namespace:HangTab.ViewModels.BottomSheets"
+    x:Class="HangTab.Views.BottomSheets.AvatarSelectBottomSheet"
     x:DataType="vm:AvatarSelectViewModel"
     HasBackdrop="True"
     HasHandle="True"
