+++ HangTab/Views/BottomSheets/AvatarSelectBottomSheet.xaml.cs
@@ -1,8 +1,8 @@
-using HangTab.ViewModels;
+using HangTab.ViewModels.BottomSheets;
 
 using The49.Maui.BottomSheet;
 
-namespace HangTab.Views;
+namespace HangTab.Views.BottomSheets;
 
 public partial class AvatarSelectBottomSheet : BottomSheet
 {
