--- HangTab/Views/Popups/DeleteBowlerPopup.xaml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<toolkit:Popup
-    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
-    xmlns:vm="clr-namespace:HangTab.ViewModels.Popups"
-    xmlns:popups="clr-namespace:HangTab.Views.Popups"
-    x:Class="HangTab.Views.Popups.DeleteBowlerPopup"
-    x:DataType="vm:DeleteBowlerPopupViewModel">
-    
-    <toolkit:Popup.Resources>
-        <Style TargetType="{x:Type popups:DeleteBowlerPopup}">
-            <Setter Property="Size" Value="300,150" />
-            <Setter Property="Color" Value="{DynamicResource PageBackgroundColor}" />
-            <Setter Property="HorizontalOptions" Value="Center" />
-            <Setter Property="VerticalOptions" Value="Start" />
-            <Setter Property="CanBeDismissedByTappingOutsideOfPopup" Value="True" />
-        </Style>
-    </toolkit:Popup.Resources>
-    
-    <VerticalStackLayout
-        Spacing="10">
-        <Label
-            Text="Delete bowler"
-            Margin="10,10,10,5"
-            Style="{StaticResource Header}" />
-        
-        <Label
-            Text="Are you sure you want to delete this bowler?"
-            TextColor="{DynamicResource TextPrimaryColor}"
-            Margin="20,5,10,5"
-            Style="{StaticResource Body}" />
-
-        <HorizontalStackLayout
-            HorizontalOptions="End"
-            Margin="10,0">
-            <Button
-                Text="No"
-                Style="{StaticResource SecondaryButton}"
-                Command="{Binding CancelCommand}" />
-            <Button
-                Text="Yes"
-                Style="{StaticResource SecondaryButton}"
-                Command="{Binding ConfirmCommand}" />
-        </HorizontalStackLayout>
-    </VerticalStackLayout>
-    
-</toolkit:Popup>
